+++ checker/src/testannotations/java/edu/umd/cs/findbugs/annotations/CheckForNull.java
@@ -1,5 +1,5 @@
 // Upstream version (this is a clean-room reimplementation of its interface):
-// http://findbugs.sourceforge.net/api/edu/umd/cs/findbugs/annotations/CheckForNull.html
+// https://findbugs.sourceforge.net/api/edu/umd/cs/findbugs/annotations/CheckForNull.html
 
 package edu.umd.cs.findbugs.annotations;