+++ framework/src/main/java/org/checkerframework/framework/source/AggregateChecker.java
@@ -136,7 +136,9 @@ public abstract class AggregateChecker extends SourceChecker {
     for (SourceChecker checker : checkers) {
       options.addAll(checker.getSupportedOptions());
     }
-    options.addAll(expandCFOptions(Arrays.asList(this.getClass()), options.toArray(new String[0])));
+    options.addAll(
+        expandCFOptions(
+            Arrays.asList(this.getClass()), options.toArray(new String[options.size()])));
     return options;
   }