+++ docs/CHANGELOG.md
@@ -1,3 +1,25 @@
+Version 3.27.1 (December1, 2022)
+---------------------------------
+
+**User-visible changes:**
+
+**Implementation details:**
+
+**Closed issues:**
+
+
+Version 3.27.0 (November 1, 2022)
+---------------------------------
+
+**User-visible changes:**
+
+The Constant Value Checker supports new annotation `@DoesNotMatchRegex`.
+
+**Closed issues:**
+
+#5238, #5360, #5362, #5387.
+
+
 Version 3.26.0 (October 3, 2022)
 --------------------------------